ref(simple-select): split large component file
