+++ packages/ng/simple-select/src/lib/panel/index.ts
@@ -1,1 +1,2 @@
 export * from './panel.component';
+export * from './panel.models';
